app/requirements.txt
Django-jazzmin==2.6.0
Django-countries==7.5
Six==1.16.0
Razorpay==1.3.0
